[Condense] Move condense settings out of experimental and defualt enable
