+++ packages/types/src/global-settings.ts
@@ -46,6 +46,7 @@ export const globalSettingsSchema = z.object({
 	alwaysAllowExecute: z.boolean().optional(),
 	allowedCommands: z.array(z.string()).optional(),
 	allowedMaxRequests: z.number().nullish(),
+	autoCondenseContext: z.boolean().optional(),
 	autoCondenseContextPercent: z.number().optional(),
 
 	browserToolEnabled: z.boolean().optional(),
@@ -131,6 +132,7 @@ export const GLOBAL_SETTINGS_KEYS = keysOf<GlobalSettings>()([
 	"alwaysAllowExecute",
 	"allowedCommands",
 	"allowedMaxRequests",
+	"autoCondenseContext",
 	"autoCondenseContextPercent",
 
 	"browserToolEnabled",
