+++ webview-ui/src/context/__tests__/ExtensionStateContext.test.tsx
@@ -203,6 +203,7 @@ describe("mergeExtensionState", () => {
 			showRooIgnoredFiles: true,
 			renderContext: "sidebar",
 			maxReadFileLine: 500,
+			autoCondenseContext: true,
 			autoCondenseContextPercent: 100,
 		}
